// types/payroll.ts
export type DeductionType = 'tax' | 'other';

export interface Employer {
  name: string;
  abn: string;
  address: string;
}

export interface Employee {
  name: string;
  id: string;
  address: string;
  classification: string;
  department: string;
}
// ...
export interface PayslipPeriod {
  index: number;
  startDate: Date;
  endDate: Date;
  paymentDate: Date;
}

export interface PayslipEarning {
  description: string;
  hours: number;
  rate: number;
  amount: number;
  ytd: number;
  type: string;
}

export interface PayslipDeduction {
  description: string;
  amount: number;
  ytd: number;
  type: DeductionType;
}

export interface PayslipLeave {
  type: string;
  accruedThisPeriod: number;
  takenThisPeriod: number;
  balance: number;
  ytdAccrued: number;
}

export interface Payslip {
  period: PayslipPeriod;
  employer: Employer;
  employee: Employee;
  earnings: PayslipEarning[];
  grossPay: number;
  ote: number;
  deductions: PayslipDeduction[];
  totalDeductions: number;
  netPay: number;
  superAmount: number;
  superYTD: number;
  leave: PayslipLeave[];
  bankAccount: string;
  paymentRef: string;
  annualRate: number;
  ytdGross: number;
  ytdNet: number;
  ytdTax: number;
}

export interface BankTransaction {
  date: Date;
  description: string;
  credit: number;
  debit: number;
  balance: number;
}